Initial commit of GNU C style string fuzzer source
Code written and tested on Ubuntu server for transliterating a buffer of characters provided on the command-line into various rotations that can be utilized as attack strings.  If the string you'd like to run it against contains whitespace, then you may have to enclose the sequence of bytes in quotes at the shell prompt!